+++ tap_intercom/client.py
@@ -327,8 +327,3 @@ class IntercomClient(object):
         if method=='POST':
             return self.post(path, **kwargs)
         return self.get(path, **kwargs)
-
-    def probe_stream(self, path, http_method='GET', **kwargs):
-        if http_method.upper() == 'POST':
-            return self.post(path, **kwargs)
-        return self.get(path, **kwargs)