+++ tests/test_intercom_discovery.py
@@ -74,10 +74,10 @@ class DiscoveryTest(IntercomBaseTest):
                     stream_properties[0].get(
                         "metadata", {self.REPLICATION_KEYS: []}).get(self.REPLICATION_KEYS, [])
                 )
-                
+
                 actual_replication_method = stream_properties[0].get(
                     "metadata", {self.REPLICATION_METHOD: None}).get(self.REPLICATION_METHOD)
-                
+
                 actual_parent_stream_id = stream_properties[0].get(
                     "metadata", {}).get(self.PARENT_TAP_STREAM_ID)
                 ##########################################################################