+++ tests/test_intercom_interrupted_sync.py
@@ -2,6 +2,7 @@ from datetime import datetime as dt
 from tap_tester import runner, connections, menagerie
 from base import IntercomBaseTest
 
+
 class intercomInterruptedSyncTest(IntercomBaseTest):
 
     def assertIsDateFormat(self, value, str_format):
@@ -20,7 +21,7 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
 
         return "intercom_interrupted_sync_test"
 
-    
+
     def test_run(self):
         """
             Scenario: A sync job is interrupted. The state is saved with `currently_syncing`.
@@ -67,7 +68,7 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
 
             connections.select_catalog_and_fields_via_metadata(conn_id, catalog, annoted_schema, \
                 additional_md=additional_md, non_selected_fields=non_selected_properties)
-            
+
         # Run sync
         record_count_by_stream_full_sync = self.run_and_verify_sync(conn_id)
         synced_records_full_sync = runner.get_records_from_target_output()
@@ -89,7 +90,7 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
                 }
             }
         }
-            
+
         # Set state for 2nd sync
         menagerie.set_state(conn_id, state)
 
@@ -131,7 +132,7 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
         # Stream level assertions
         for stream in expected_streams:
             with self.subTest(stream=stream):
-                
+
                 # Gather Actual results
                 full_records = [message['data'] for message in synced_records_full_sync.get(stream, {}).get('messages', [])]
                 full_record_count = record_count_by_stream_full_sync.get(stream, 0)
@@ -147,7 +148,7 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
                     final_state_bookmark_datetime = dt.strptime(final_stream_bookmark, "%Y-%m-%dT%H:%M:%S.%fZ")
                     full_sync_bookmark = full_sync_state['bookmarks'][stream].get("updated_at")
                     full_sync_bookmark_datetime = dt.strptime(full_sync_bookmark, "%Y-%m-%dT%H:%M:%S.%fZ")
-                    
+
                     # Verify final bookmark saved matches formatting standards for resuming sync
                     self.assertIsNotNone(final_stream_bookmark)
                     self.assertIsInstance(final_stream_bookmark, str)
@@ -155,7 +156,7 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
 
 
                     if stream == state['currently_syncing']:
-                       
+
                         # Check if the interrupted stream has a bookmark written for it
                         interrupted_stream_bookmark = state['bookmarks'][stream].get("updated_at")
                         interrupted_bookmark_datetime = dt.strptime(interrupted_stream_bookmark, "%Y-%m-%dT%H:%M:%S.%fZ")
@@ -168,7 +169,7 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
                             self.assertGreaterEqual(rec_time, interrupted_bookmark_datetime)
 
                             # As live data is received and bookmark is getting updated in between the sync, some records might not be present.
-                            if full_sync_bookmark_datetime > final_state_bookmark_datetime: 
+                            if full_sync_bookmark_datetime > final_state_bookmark_datetime:
                                 self.assertIn(record, full_records, msg='Incremental table record in interrupted sync not found in full sync')
 
                         # Record count for all streams of interrupted sync match expectations
@@ -188,10 +189,10 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
                             synced_stream_datetime = dt.strptime(synced_stream_bookmark, "%Y-%m-%dT%H:%M:%S.%fZ")
                         else:
                             synced_stream_datetime = start_date_datetime
-                        
+
                         # Verify we replicated some records for the non-interrupted streams
                         self.assertGreater(interrupted_record_count, 0)
-                        
+
                         # - Verify resuming sync only replicates records with replication key values greater or equal to
                         #       the state for streams that were replicated during the interrupted sync.
                         # - Verify resuming sync replicates all records that were found in the full sync (un-interupted)
@@ -200,10 +201,10 @@ class intercomInterruptedSyncTest(IntercomBaseTest):
                             self.assertGreaterEqual(rec_time, synced_stream_datetime)
 
                             self.assertIn(record, full_records, msg='Unexpected record replicated in resuming sync.')
-                        
+
                         # Verify we replicated all the records from 1st sync for the streams
                         #       that are left to sync (ie. streams without bookmark in the state)
-                        if stream not in state["bookmarks"].keys(): 
+                        if stream not in state["bookmarks"].keys():
                             for record in full_records:
                                 self.assertIn(record, interrupted_records, msg='Record missing from resuming sync.' )